+++ admin/partials/whats-playing.php
@@ -3,6 +3,7 @@
 	<h2>Whats Playing</h2>
 
 	<?php
+		settings_errors();
 		settings_fields( 'whats_playing' );
 		do_settings_sections( 'whats_playing' );
 		submit_button('Authenticate');