+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+## [v1.0.1](https://github.com/peterjohnhunt/whats-playing/compare/v1.0.1...v1.0.1) | 2017-03-29
+- Added Refresh Token Functionality
+- Added Admin Notices
+- Refactored Settings Storage
+- Improved Settings Validation
+- Added Graceful Error Handling
+- Additional Minor Fixes
+- Readme Updated
 ## [v1.0.0](https://github.com/peterjohnhunt/whats-playing/releases/tag/v1.0.0) | 2017-03-29
 - Initial v1.0.0 Release